Renamed DateList to DatesList; js_only.
