+++ assets/prototype/domain/eventEditor/datetimes/datesList/filterBar/DatesListFilterBar.js
