+++ assets/prototype/domain/eventEditor/datetimes/datesList/filterBar/controls/DatesSortedByControl.js
